feat: clean up page, add paypal
